DisasterEvent親クラスを作成
