+++ app/app/Console/Commands/ExecuteTurn.php
@@ -2,6 +2,7 @@
 
 namespace App\Console\Commands;
 
+use App\Entity\Event\Disaster\Disaster;
 use App\Entity\Event\ForeignIsland\ForeignIslandEvent;
 use App\Entity\Log\AbandonmentLog;
 use App\Entity\Log\AbortInvalidIslandLog;
@@ -77,7 +78,7 @@ class ExecuteTurn extends Command
                 $prevStatusList = new Collection();
                 $logsList = new Collection();
                 $foreignIslandTargetedPlans = new Collection();
-                $foreignIslandOccurEvents = new Collection();
+                $foreignIslandEvents = new Collection();
 
                 /** @var Island $island */
                 foreach ($islands as $island) {
@@ -166,13 +167,13 @@ class ExecuteTurn extends Command
                     $logs = $logsList->get($island->id);
 
                     // 災害
-                    $occurDisasterResult = $terrain->occurDisaster($island, $status, $turn);
-                    $status = $occurDisasterResult->getStatus();
-                    $logs->merge($occurDisasterResult->getLogs());
+                    $disasterResult = Disaster::occur($island, $terrain, $status, $turn);
+                    $status = $disasterResult->getStatus();
+                    $logs->merge($disasterResult->getLogs());
 
                     // セル処理
                     // FIXME: 本来災害はセル処理の後だが、隕石→湖判定の順番を考慮し逆にしている
-                    $passTurnResult = $terrain->passTurn($island, $status, $turn, $foreignIslandOccurEvents);
+                    $passTurnResult = $terrain->passTurn($island, $status, $turn, $foreignIslandEvents);
                     $status = $passTurnResult->getStatus();
                     $logs->merge($passTurnResult->getLogs());
 
@@ -205,7 +206,7 @@ class ExecuteTurn extends Command
 
                 // セル処理によって発生した計画の実行
                 /** @var ForeignIslandEvent $plan */
-                foreach ($foreignIslandOccurEvents as $plan) {
+                foreach ($foreignIslandEvents as $plan) {
                     /** @var Island $toIsland */
                     $toIsland = $islands->find($plan->getToIsland());
                     /** @var Island $fromIsland */
