+++ app/app/Entity/Cell/Cell.php
@@ -154,7 +154,7 @@ abstract class Cell
         return new (CellTypeConst::getClassByType($type))(...get_object_vars($data));
     }
 
-    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandOccurEvents): PassTurnResult
+    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandEvents): PassTurnResult
     {
         return new PassTurnResult($terrain, $status, Logs::create());
     }
