+++ app/app/Entity/Cell/MissileBase/MissileBase.php
@@ -94,7 +94,7 @@ class MissileBase extends Cell implements IMissileFireable
         return '('. $this->point->x . ',' . $this->point->y .') ' . Forest::NAME;
     }
 
-    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandOccurEvents): PassTurnResult
+    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandEvents): PassTurnResult
     {
         $cells = $terrain->getAroundCells($this->point);
         $seasideCells = $cells->reject(function ($cell) { return $cell::ATTRIBUTE[CellTypeConst::IS_LAND]; });
