+++ app/app/Entity/Cell/Monster/Kujira.php
@@ -79,14 +79,14 @@ class Kujira extends Monster
         return self::DEFAULT_MOVE_TIMES;
     }
 
-    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandOccurEvents): PassTurnResult
+    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandEvents): PassTurnResult
     {
         if (($turn->turn + 1) % 3 === 0) {
             $this->isMetalized = true;
             return new PassTurnResult($terrain, $status, Logs::create());
         } else {
             $this->isMetalized = false;
-            return parent::passTurn($island, $terrain, $status, $turn, $foreignIslandOccurEvents);
+            return parent::passTurn($island, $terrain, $status, $turn, $foreignIslandEvents);
         }
     }
 }
