+++ app/app/Entity/Cell/Monster/Monster.php
@@ -99,7 +99,7 @@ abstract class Monster extends Cell
         return false;
     }
 
-    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandOccurEvents): PassTurnResult
+    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandEvents): PassTurnResult
     {
         if ($this->remainMoveTimes <= 0) {
             return new PassTurnResult($terrain, $status, Logs::create());
@@ -137,7 +137,7 @@ abstract class Monster extends Cell
         $monster->point = $moveTarget->point;
         // 移動先でさらに動く場合の操作をするため再帰呼び出しをしている
         $terrain->setCell($monster->getPoint(), $monster);
-$passTurnResult = $terrain->getCell($monster->getPoint())->passTurn($island, $terrain, $status, $turn, $foreignIslandOccurEvents);
+$passTurnResult = $terrain->getCell($monster->getPoint())->passTurn($island, $terrain, $status, $turn, $foreignIslandEvents);
 
         $terrain = $passTurnResult->getTerrain();
         $status = $passTurnResult->getStatus();
