+++ app/app/Entity/Cell/Park/MonumentOfMining.php
@@ -33,7 +33,7 @@ class MonumentOfMining extends Park
         return true;
     }
 
-    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandOccurEvents): PassTurnResult
+    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandEvents): PassTurnResult
     {
         $status->setDevelopmentPoints($status->getDevelopmentPoints() + self::PRODUCT_DEVELOPMENT_POINTS);
         return new PassTurnResult($terrain, $status, Logs::create());
