+++ app/app/Entity/Cell/Park/MonumentOfWar.php
@@ -39,7 +39,7 @@ class MonumentOfWar extends Park
         return false;
     }
 
-    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandOccurEvents): PassTurnResult
+    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandEvents): PassTurnResult
     {
         $status->setDevelopmentPoints($status->getDevelopmentPoints() + self::PRODUCT_DEVELOPMENT_POINTS);
         return new PassTurnResult($terrain, $status, Logs::create());
