+++ app/app/Entity/Cell/Ship/Ship.php
@@ -80,7 +80,7 @@ abstract class Ship extends Cell
         return $terrain;
     }
 
-    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandOccurEvents): PassTurnResult
+    public function passTurn(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandEvents): PassTurnResult
     {
         $logs = Logs::create();
