--- app/app/Entity/Event/Disaster/DisasterConst.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Entity\Event\Disaster;
-
-class DisasterConst
-{
-    public const DISASTERS = [
-        Fire::class,
-        Earthquake::class,
-        Tsunami::class,
-        VolcanicEruption::class,
-        Typhoon::class,
-        Meteorite::class,
-        HugeMeteorite::class,
-        Riot::class,
-        LandSubsidence::class,
-        AppearanceMonster::class,
-        PirateInvasion::class,
-    ];
-}
